Fixed a bug in the active connection sniffing routine
Active connection sniffing was limited to 3 sniffers only, whereas it could use
as many sniffers as we want to recover a connection's channel map. Limit has been
increased to 999 as a dirty fix, but I'm pretty sure this could be done a better
way. I did it in a urge, so don't blame me (but you should).